feat: refactor DTO imports and create GetDoctorQuery class for improved doctor data handling
