+++ src/doctors/doctors.service.ts
@@ -7,7 +7,8 @@ import { IDoctorsRepository } from './doctors.repository.interface';
 import { DoctorEntity } from './entities/doctor.entity';
 import { ISpecializationsService } from 'src/specializations/specializations.service.interface';
 import { ILogger } from 'src/common/interfaces/logger.interface';
-import { DoctorDto, GetDoctorQuery } from './dto/doctor.dto';
+import { DoctorDto } from './dto/doctor.dto';
+import { GetDoctorQuery } from './dto/get-doctor-query.dto';
 import { IAppointmentsRepository } from 'src/appointments/appointments.repository.interface';
 import { CreateDoctorsRatingDto } from './dto/create-doctor-rating.dto';
 import { DoctorsRatingDto } from './dto/doctor-rating.dto';
